Only move dist config file once.

+++ mangoszero/launch_mangosd.sh
@@ -1,6 +1,8 @@
 #!/bin/bash -e
 
-mv etc/mangosd.conf.dist etc/mangosd.conf
+if [ -f etc/mangosd.conf.dist ]; then
+  mv etc/mangosd.conf.dist etc/mangosd.conf
+fi
 
 sed -i "s/^LoginDatabaseInfo *=.*$/LoginDatabaseInfo = $LOGIN_DATABASE_INFO/" etc/mangosd.conf
 sed -i "s/^WorldDatabaseInfo *=.*$/WorldDatabaseInfo = $WORLD_DATABASE_INFO/" etc/mangosd.conf
@@ -10,5 +12,7 @@ if [ -f /mangosconf/ahbot.conf ]; then
   echo "/mangosdconf/ahbot.conf is being used"
   AHCONFIG="-a /mangosconf/ahbot.conf"
 fi
+
+set -x
 
 ./bin/mangosd -c etc/mangosd.conf ${AHCONFIG}
